Demo Game
Demo Game and clean up of unity project (multiple imports of SimpleJson package)
