+++ Assets/PlayroomKit.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 58f0dab07b48d014ab3a551354cd9b29
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
