+++ Assets/Resources/Player.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c14e0b3aef64ed84ebfdf32c2def9cac
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
