+++ Packages/packages-lock.json
@@ -1,12 +1,5 @@
 {
   "dependencies": {
-    "codes.aureola.simplejson": {
-      "version": "https://github.com/aureola-codes/SimpleJSON-Unity.git",
-      "depth": 0,
-      "source": "git",
-      "dependencies": {},
-      "hash": "3b41a4c4f63b010cac95fbef69c216c38796e33d"
-    },
     "com.unity.2d.animation": {
       "version": "7.0.6",
       "depth": 1,
